requirements.txt
requests>=2.31.0
PyMuPDF>=1.23.0
pyyaml>=6.0
pydantic>=2.5.0
httpx>=0.26.0
b2sdk>=2.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
